Added clandmark facial landmarking

+++ app/src/main/java/com/sci2015fair/opencv/Classify.java
@@ -16,13 +16,20 @@ import android.util.Log;
 import android.widget.Toast;
 
 import com.sci2015fair.R;
+import com.sci2015fair.service.SD;
 
 import org.opencv.android.BaseLoaderCallback;
 import org.opencv.android.LoaderCallbackInterface;
 import org.opencv.android.OpenCVLoader;
 import org.opencv.android.Utils;
+import org.opencv.core.CvType;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfRect;
+import org.opencv.core.Point;
+import org.opencv.core.Rect;
+import org.opencv.core.Scalar;
+import org.opencv.core.Size;
+import org.opencv.imgproc.Imgproc;
 import org.opencv.objdetect.CascadeClassifier;
 
 import java.io.File;
@@ -30,16 +37,18 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 
-import libsvm.*;
-
 /**
  * Service to classify photos using OpenCV
  */
 public class Classify extends Service {
     private final String TAG = "OCV";
-    private Looper mServiceLooper;
     private ServiceHandler mServiceHandler;
+    private String imgPath = null;
 
+    public native int[] runLandmarks(int cols, int rows, int[] bbox, byte[] data, String path);
+    static {
+        System.loadLibrary("clandmark");
+    }
     /**
      * OpenCV Callback
      */
@@ -50,7 +59,6 @@ public class Classify extends Service {
                 case LoaderCallbackInterface.SUCCESS:
                 {
                     Log.i(TAG, "OpenCV loaded successfully");
-                    //mOpenCvCameraView.enableView();
                 } break;
                 default:
                 {
@@ -77,13 +85,10 @@ public class Classify extends Service {
                 mLoaderCallback.onManagerConnected(LoaderCallbackInterface.SUCCESS);
             }
 
-            /*
-            File root = new File(Environment.getExternalStoragePublicDirectory(
-                    Environment.DIRECTORY_PICTURES), "2015-ScienceFair"); //get root of image directory
-
             Mat image = new Mat();
+            File originalImage = new File(imgPath);
             try {
-                FileInputStream fis = new FileInputStream(new File(root, "IMG_20151005_141009b.jpg")); //get image
+                FileInputStream fis = new FileInputStream(originalImage); //get image
                 Bitmap bitmap = BitmapFactory.decodeStream(fis);
                 Utils.bitmapToMat(bitmap, image); //convert to Mat
                 fis.close();
@@ -92,21 +97,26 @@ public class Classify extends Service {
             }
 
             // load cascade file from application resources
-            InputStream is = getResources().openRawResource(R.raw.haar_leye);
+            // needs to write resource to file to get a valid path
+            InputStream is = getResources().openRawResource(R.raw.haarcascade_frontalface_alt);
             File cascadeDir = getDir("cascade", Context.MODE_PRIVATE);
-            File mCascadeFile = new File(cascadeDir, "haar_leye.xml");
-            FileOutputStream os = null;
-            try {
-                os = new FileOutputStream(mCascadeFile);
-                byte[] buffer = new byte[4096];
-                int bytesRead;
-                while ((bytesRead = is.read(buffer)) != -1) {
-                    os.write(buffer, 0, bytesRead);
+            File mCascadeFile = new File(cascadeDir, "haarcascade_frontalface_alt.xml");
+            if(!mCascadeFile.exists()){
+                FileOutputStream os = null;
+                try {
+                    os = new FileOutputStream(mCascadeFile);
+                    byte[] buffer = new byte[4096];
+                    int bytesRead;
+                    while ((bytesRead = is.read(buffer)) != -1) {
+                        os.write(buffer, 0, bytesRead);
+                    }
+                    is.close();
+                    os.close();
+                } catch (Exception e) {
+                    e.printStackTrace();
                 }
-                is.close();
-                os.close();
-            } catch (Exception e) {
-                e.printStackTrace();
+            } else {
+                Log.i(TAG, "Cascade file exists.");
             }
 
             CascadeClassifier faceDetector = new CascadeClassifier(mCascadeFile.getAbsolutePath());
@@ -115,23 +125,86 @@ public class Classify extends Service {
                 faceDetector = null;
             } else
                 Log.i(TAG, "Loaded cascade classifier from " + mCascadeFile.getAbsolutePath());
+            //mCascadeFile.delete(); //delete after it is loaded
 
-            cascadeDir.delete(); //delete temp cascade XML after it is loaded
-            Log.d("ROWS", String.format("R: %d    C: %d", image.rows(), image.cols()));
+            Log.d(TAG, String.format("Rows: %d   Cols: %d", image.rows(), image.cols()));
 
-            MatOfRect faceDectections = new MatOfRect(); //array to hold info on detected faces
-            faceDetector.detectMultiScale(image, faceDectections); //detect faces
+            MatOfRect faceDetections = new MatOfRect(); //array to hold info on detected faces
+            faceDetector.detectMultiScale(image, faceDetections); //detect faces
 
-            Log.d("FACE",String.format("Detected %d faces", faceDectections.toArray().length));
+            Log.d(TAG, String.format("Detected %d faces", faceDetections.toArray().length));
+
+            Mat temp = image.clone();
+            Imgproc.cvtColor(image,temp,Imgproc.COLOR_BGR2GRAY);
+            temp.convertTo(temp, CvType.CV_8U); //convert to grayscale
+            byte[] data = new byte[temp.rows()*temp.cols()]; //save into continuous array
+            for(int y = 0; y<temp.rows(); y++){
+                for(int x = 0; x<temp.cols(); x++){
+                    data[y*temp.cols()+x] = (byte)temp.get(y,x)[0];
+                }
+            }
+
+            //load flandmark model
+            InputStream ins = getResources().openRawResource(R.raw.frontalaflw);
+            File landDir = getDir("flandmark", Context.MODE_PRIVATE);
+            File land = new File(landDir, "flandmark_model.xml");
+            if(!land.exists()){
+                FileOutputStream fos = null;
+                try {
+                    fos = new FileOutputStream(land);
+                    byte[] buffer = new byte[4096];
+                    int bytesRead;
+                    while ((bytesRead = ins.read(buffer)) != -1) {
+                        fos.write(buffer, 0, bytesRead);
+                    }
+                    ins.close();
+                    fos.close();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            } else {
+                Log.i(TAG, "Landmark file exists.");
+            }
+
+            String landPath = land.getAbsolutePath(); //path to model
+            int[] out; //data from clandmark
+
+            for(Rect rect : faceDetections.toArray()){
+                rect.height *= 1.15; //increase height slightly to compensate for smaller window
+                int[] bbox = new int[8];
+                bbox[0] = rect.x;
+                bbox[1] = rect.y;
+                bbox[2] = rect.x+rect.width;
+                bbox[3] = rect.y;
+                bbox[4] = rect.x+rect.width;
+                bbox[5] = rect.y+rect.height;
+                bbox[6] = rect.x;
+                bbox[7] = rect.y+rect.height;
+
+                for(int i = 0; i<8; i+=2){ //print out bounding box
+                    Log.d(TAG, bbox[i] + ", " + bbox[i+1]);
+                }
+                Imgproc.rectangle(image, new Point(bbox[0], bbox[1]), new Point(bbox[4], bbox[5]),
+                        new Scalar(255, 0, 0)); //draw bounding box
+
+                out = runLandmarks(image.width(),image.height(),bbox,data,landPath); //call clandmark code
+                for(int i = 0; i<out.length; i+=2){ //display points
+                    Log.d(TAG, out[i] + ", " + out[i + 1]);
+                    Imgproc.rectangle(image, new Point(out[i], out[i+1]), new Point(out[i] + 2, out[i+1] + 2),
+                            new Scalar(0, 255, 0));
+                }
+                break; //only analyze the first face
+            }
+            Bitmap bmp = Bitmap.createBitmap(image.width(),image.height(),Bitmap.Config.ARGB_8888); //convert back
+            Utils.matToBitmap(image, bmp);
+            SD.saveImage(bmp,true);
+            //land.delete(); //delete after it is loaded
+            // originalImage.delete();//remove old image
             // Stop the service using the startId, so that we don't stop
             // the service in the middle of handling another job
-            */
-
-
             stopSelf(msg.arg1);
         }
     }
-
 
     @Override
     public void onCreate() {
@@ -144,13 +217,17 @@ public class Classify extends Service {
         thread.start();
 
         // Get the HandlerThread's Looper and use it for our Handler
-        mServiceLooper = thread.getLooper();
+        Looper mServiceLooper = thread.getLooper();
         mServiceHandler = new ServiceHandler(mServiceLooper);
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         Toast.makeText(this, "OCV Starting", Toast.LENGTH_SHORT).show();
+
+        if (intent.hasExtra("filepath")) {
+            imgPath = intent.getStringExtra("filepath");
+        }
 
         // For each start request, send a message to start a job and deliver the
         // start ID so we know which request we're stopping when we finish the job
@@ -159,9 +236,8 @@ public class Classify extends Service {
         mServiceHandler.sendMessage(msg);
 
         // If we get killed, after returning from here, restart
-        return START_STICKY;
+        return START_REDELIVER_INTENT;
     }
-
 
     @Override
     public IBinder onBind(Intent intent) {
